fix: resolve all clippy lints across workspace
- vo-codegen: too_many_arguments allows, iter().cloned().collect()→to_vec(),
  repeat_n, expect(format!())→panic!(), let_and_return, redundant u16/u32 casts,
  clone on Copy types, useless format!/vec!, match→if let, map_or→is_some_and/
  is_none_or, unnecessary if let→flatten, loop index→enumerate, identical if blocks
- vo-engine: sort_by→sort_by_key, format!→to_string, items_after_test_module,
  clone→slice::from_ref
- vo-web/runtime-wasm: thread_local const initializers, match→if, Option::map,
  strip_prefix, redundant refs
- cmd/vo: starts_with+slice→strip_prefix
- vo-ext: add Default impl for HostBridge
- vo-runtime: items_after_test_module, while-let→for, approx PI→f64::consts::PI
- vo-vm: items_after_test_module
- vo-analysis: unused var, cargo fmt fixes

All: cargo clippy -D warnings ✓  cargo fmt --check ✓  ./d.py test jit --release 1012 passed

+++ lang/crates/vo-codegen/src/stmt/defer_go.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::too_many_arguments)]
 //! Defer and go statement compilation.
 
 use vo_runtime::SlotType;
@@ -656,7 +657,7 @@ fn compile_defer_pkg_func_call(
     // Get return slot count from function signature
     let func_type = info.expr_type(call_expr.func.id);
     let sig = info.as_signature(func_type);
-    let ret_slots = info.type_slot_count(sig.results()) as u16;
+    let ret_slots = info.type_slot_count(sig.results());
 
     let wrapper_id = crate::wrapper::generate_defer_extern_wrapper(
         ctx,
@@ -1177,7 +1178,7 @@ fn compile_go_pkg_func_call(
     let extern_name = crate::expr::call::get_extern_name(sel, info)?;
     let func_type = info.expr_type(call_expr.func.id);
     let sig = info.as_signature(func_type);
-    let ret_slots = info.type_slot_count(sig.results()) as u16;
+    let ret_slots = info.type_slot_count(sig.results());
     let wrapper_id = crate::wrapper::generate_defer_extern_wrapper(
         ctx,
         &extern_name,
